a complex makefile

+++ me.h
@@ -1,1 +1,1 @@
-float get_age();
+int get_age();
